working on follower

+++ raft-protocol/src/main/java/com/josiahebhomenye/raft/RequestVoteReply.java
@@ -8,6 +8,6 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RequestVoteReply {
-    int term;
+    long term;
     boolean voteGranted;
 }
